Test: (BRD-133) 게시글 단건 조회 API 명세 문서화(RestDocs)
