+++ board-system-app/src/docs/asciidoc/article-api.adoc
@@ -21,6 +21,22 @@ include::{snippets}/article-create-success/http-request.adoc[]
 include::{snippets}/article-create-success/http-response.adoc[]
 
 === 게시글 단건 조회
+==== 기본정보
+- 메서드 : GET
+- URL : `/api/v1/articles/{articleId}`
+- 권한 : 누구나
+
+게시글을 단건 조회합니다.
+
+==== 요청
+include::{snippets}/article-read-success/path-parameters.adoc[]
+
+==== 응답
+include::{snippets}/article-read-success/response-fields.adoc[]
+
+==== 예시
+include::{snippets}/article-read-success/http-request.adoc[]
+include::{snippets}/article-read-success/http-response.adoc[]
 
 === 게시글 목록 조회(페이지)
 ==== 기본정보
